# frontend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/applications/hr/jobs/{jobId}": {
      "get": {
        "description": "Returns all applications for a job including:\n- HR decision view\n- Technical evaluation\n- Candidate info\n\nSupports filtering.",
        "operationId": "ApplicantsController_getJobApplications",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "job_123",
              "type": "string"
            }
          },
          {
            "name": "fitTier",
            "required": false,
            "in": "query",
            "description": "Filter by fit tier",
            "schema": {
              "enum": [
                "STRONG",
                "PROBE",
                "PASS"
              ],
              "type": "string"
            }
          },
          {
            "name": "minScore",
            "required": false,
            "in": "query",
            "description": "Minimum role fit score",
            "schema": {
              "example": 70,
              "type": "number"
            }
          },
          {
            "name": "pipelineStage",
            "required": false,
            "in": "query",
            "description": "Filter by pipeline stage",
            "schema": {
              "enum": [
                "APPLIED",
                "REVIEWED",
                "SHORTLISTED",
                "INTERVIEW_HR",
                "INTERVIEW_TECHNICAL",
                "INTERVIEW_FINAL",
                "OFFER",
                "HIRED",
                "REJECTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Applications retrieved successfully"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "List job applications (HR)",
        "tags": [
          "Applications"
        ]
      }
    },
...
    "/jobs": {
      "get": {
        "description": "Returns a list of published jobs with filtering and pagination.",
        "operationId": "JobsController_getPublicJobs",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in job title and description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleType",
            "required": false,
            "in": "query",
            "description": "Role type required for the job",
            "schema": {
              "type": "string",
              "enum": [
                "BACKEND",
                "FRONTEND",
                "FULLSTACK",
                "INFRASTRUCTURE",
                "DATA_ML",
                "SMART_CONTRACT",
                "WEB3_BACKEND",
                "WEB3_FRONTEND",
                "WEB3_FULLSTACK",
                "DEFI_PROTOCOL",
                "SECURITY_WEB3",
                "SECURITY",
                "GENERALIST"
              ]
            }
          },
          {
            "name": "seniority",
            "required": false,
            "in": "query",
            "description": "Seniority level required",
            "schema": {
              "type": "string",
              "enum": [
                "JUNIOR",
                "MID",
                "SENIOR",
                "LEAD"
              ]
            }
          },
          {
            "name": "isWeb3",
            "required": false,
            "in": "query",
            "description": "Filter Web3 jobs",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (max: 50)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Browse open jobs (Public)",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/me": {
      "get": {
        "description": "Returns all job posts owned by the authenticated company, ordered by newest first.\n\nUseful for dashboards and job management views.",
        "operationId": "JobsController_getMyJobs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of jobs",
            "schema": {
              "example": {
                "success": true,
                "data": [
                  {
                    "id": "job_1",
                    "title": "Backend Engineer",
                    "status": "ACTIVE"
                  }
                ]
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all jobs created by the authenticated company",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}": {
      "get": {
        "description": "Returns detailed information about a published job.",
        "operationId": "JobsController_getPublicJobById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get job details (Public)",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/draft": {
      "post": {
        "description": "Creates a job in DRAFT status for the authenticated company.\n\nUse this endpoint when a company wants to start drafting a job listing before publishing it.",
        "operationId": "JobsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "schema": {
              "example": {
                "success": true,
                "message": "Job created successfully",
                "data": {
                  "id": "cma9x1k2p0000qwert123",
                  "title": "Senior Backend Engineer",
                  "description": "We are hiring...",
                  "status": "DRAFT",
                  "companyId": "company_123",
                  "createdAt": "2026-04-27T10:00:00.000Z",
                  "updatedAt": "2026-04-27T10:00:00.000Z"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid job payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new job post draft",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/parse-jd": {
      "post": {
        "description": "Uses AI to extract structured requirements (skills, seniority, weights).\n\nFrontend should use this for preview before confirming requirements.",
        "operationId": "JobsController_parseJd",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job ID",
            "schema": {
              "example": "cma9x1k2p0000qwert123",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseJobDescriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parsed job requirements preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseJdResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Parse job description into structured requirements",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/confirm-requirements": {
      "post": {
        "description": "Persists AI-parsed requirements into the job.\n\nThis step finalizes the structured scoring configuration.",
        "operationId": "JobsController_confirmRequirements",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job ID",
            "schema": {
              "example": "cma9x1k2p0000qwert123",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParsedJobRequirementsSwaggerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Requirements confirmed and saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm parsed job requirements",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/publish": {
      "post": {
        "description": "Changes job status from DRAFT to ACTIVE.\n\nOnce published, the job becomes visible to candidates.",
        "operationId": "JobsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job ID",
            "schema": {
              "example": "cma9x1k2p0000qwert123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Job not found or does not belong to user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a job",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/close": {
      "post": {
        "description": "Marks the job as CLOSED.\n\nClosed jobs are no longer visible or accepting applications.",
        "operationId": "JobsController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job ID",
            "schema": {
              "example": "cma9x1k2p0000qwert123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close a job",
        "tags": [
          "Jobs"
        ]
      }
    },
...
  },
  "info": {
    "title": "a16zero Employer API",
    "description": "Backend MVP for Employer platform features and Account Abstraction auth verification.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "internal-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-key",
        "description": "Internal API key"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "JobResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cma9x1k2p0000qwert123"
          },
          "title": {
            "type": "string",
            "example": "Senior Backend Engineer"
          },
          "description": {
            "type": "string",
            "example": "We are looking for a NestJS expert to build scalable APIs..."
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_PAYMENT",
              "ACTIVE",
              "CLOSED_PENDING",
              "CLOSED"
            ],
            "example": "ACTIVE",
            "description": "Job status"
          },
          "companyId": {
            "type": "string",
            "example": "company_123"
          },
          "roleType": {
            "type": "string",
            "enum": [
              "BACKEND",
              "FRONTEND",
              "FULLSTACK",
              "INFRASTRUCTURE",
              "DATA_ML",
              "SMART_CONTRACT",
              "WEB3_BACKEND",
              "WEB3_FRONTEND",
              "WEB3_FULLSTACK",
              "DEFI_PROTOCOL",
              "SECURITY_WEB3",
              "SECURITY",
              "GENERALIST"
            ],
            "example": "BACKEND"
          },
          "seniorityLevel": {
            "type": "string",
            "enum": [
              "JUNIOR",
              "MID",
              "SENIOR",
              "LEAD"
            ],
            "example": "SENIOR"
          },
          "isWeb3Role": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-04-27T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-04-27T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "status",
          "companyId",
          "roleType",
          "seniorityLevel",
          "isWeb3Role",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateJobDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Senior NestJS Engineer"
          },
          "description": {
            "type": "string",
            "example": "We are looking for a senior engineer to join our team."
          },
          "location": {
            "type": "string",
            "example": "Remote"
          },
          "employmentType": {
            "type": "string",
            "example": "Full-time"
          },
          "bonusAmount": {
            "type": "number",
            "example": 1000
          },
          "currency": {
            "type": "string",
            "example": "USD"
          }
        },
        "required": [
          "title",
          "description"
        ]
      },
      "ParseJobDescriptionDto": {
        "type": "object",
        "properties": {
          "jdText": {
            "type": "string",
            "example": "We are looking for a Senior Backend Engineer proficient in NestJS and PostgreSQL..."
          }
        },
        "required": [
          "jdText"
        ]
      },
      "ParseJdResponseDto": {
        "type": "object",
        "properties": {
          "parsed": {
            "type": "object",
            "description": "AI parsed structured requirements",
            "example": {
              "skills": [
                "Node.js",
                "NestJS",
                "PostgreSQL"
              ],
              "seniority": "SENIOR",
              "parserConfidence": 0.82
            }
          },
          "requiresReview": {
            "type": "boolean",
            "description": "Whether manual review is recommended",
            "example": false
          },
          "diff": {
            "type": "object",
            "description": "Diff summary of extracted requirements",
            "example": {
              "added": [
                "GraphQL"
              ],
              "removed": [],
              "changedWeights": [
                "ownershipWeight"
              ]
            }
          }
        },
        "required": [
          "parsed",
          "requiresReview",
          "diff"
        ]
      },
      "ParsedJobRequirementsSwaggerDto": {
        "type": "object",
        "properties": {
          "requiredSkills": {
            "description": "Technologies required for the role",
            "example": [
              "Node.js",
              "NestJS",
              "PostgreSQL"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requiredRoleType": {
            "type": "string",
            "enum": [
              "BACKEND",
              "FRONTEND",
              "FULLSTACK",
              "INFRASTRUCTURE",
              "DATA_ML",
              "SMART_CONTRACT",
              "WEB3_BACKEND",
              "WEB3_FRONTEND",
              "WEB3_FULLSTACK",
              "DEFI_PROTOCOL",
              "SECURITY_WEB3",
              "SECURITY",
              "GENERALIST"
            ],
            "example": "BACKEND",
            "description": "Primary role type expected for the candidate"
          },
          "seniorityLevel": {
            "type": "string",
            "enum": [
              "JUNIOR",
              "MID",
              "SENIOR",
              "LEAD"
            ],
            "example": "SENIOR",
            "description": "Expected seniority level"
          },
          "collaborationWeight": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "example": "HIGH",
            "description": "Importance of collaboration skills"
          },
          "ownershipWeight": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "example": "MEDIUM",
            "description": "Importance of ownership/autonomy"
          },
          "innovationWeight": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "example": "LOW",
            "description": "Importance of innovation and creativity"
          },
          "isWeb3Role": {
            "type": "boolean",
            "example": true,
            "description": "Whether this role is Web3-related"
          },
          "parserConfidence": {
            "type": "number",
            "example": 0.82,
            "description": "Confidence score from the AI parser (0 to 1). Values below ~0.75 should be reviewed manually.",
            "minimum": 0,
            "maximum": 1
          }
        },
        "required": [
          "requiredSkills",
          "requiredRoleType",
          "seniorityLevel",
          "collaborationWeight",
          "ownershipWeight",
          "innovationWeight",
          "isWeb3Role",
          "parserConfidence"
        ]
      },
...
    }
  }
}